🚚(frontend) reorganize apps components
Our frontend apps components didn't match the structure of our main
application.
