+++ src/frontend/apps/.cookiecutter/{{cookiecutter.app_name}}/components/Dashboard{{cookiecutter.model}}/index.tsx
