+++ src/frontend/apps/.cookiecutter/{{cookiecutter.app_name}}/components/SelectContentTab/index.spec.tsx
