+++ src/frontend/apps/bbb/components/DashboardClassroomError/index.spec.tsx
