+++ src/frontend/apps/bbb/components/DashboardClassroomInfos/index.spec.tsx
