+++ src/frontend/apps/bbb/components/DashboardClassroomJoin/index.spec.tsx
