+++ src/frontend/apps/bbb/components/DashboardClassroomLayout/index.spec.tsx
