+++ src/frontend/apps/bbb/components/DashboardClassroomLayout/index.tsx
