+++ src/frontend/apps/markdown/components/MdxRenderer/__tests__/markdownLatex.md
