+++ src/frontend/apps/markdown/components/MdxRenderer/__tests__/markdownMermaid.md
