+++ src/frontend/apps/markdown/components/MdxRenderer/index.spec.tsx
@@ -17,7 +17,7 @@ jest.mock('data/stores/useAppConfig', () => ({
   },
 }));
 
-jest.mock('apps/markdown/MdxRenderer/constants', () => ({
+jest.mock('apps/markdown/components/MdxRenderer/constants', () => ({
   debouncingTime: 0, // override the debouncing time to make tests twice faster
 }));
