+++ src/frontend/apps/markdown/components/MdxRenderer/remarkLatexPlugin.tsx
