+++ src/frontend/apps/markdown/components/MdxRenderer/remarkMermaidPlugin.tsx
